move shader to shadernode
