+++ examples/phong_light_model.py
@@ -43,7 +43,7 @@ geometry.setAttribute('uv', three.Float32BufferAttribute(uvs, 2))
 geometry.setIndex(pmd_file.indices)
 # geometry.computeVertexNormals()
 
-canvas = WgpuCanvas(size=(640, 480), title="Phong Light Model")
+canvas = WgpuCanvas(size=(640, 480),  max_fps=60, title="Phong Light Model")
 render = three.WgpuRenderer(canvas, parameters={'antialias': True})
 
 render.init()
