+++ three/nodes/display/normal_map_node.py
@@ -2,8 +2,8 @@ from ..core.temp_node import TempNode
 from ..core.node_builder import NodeBuilder
 from ..accessors.model_node import ModelNode
 
-from ..shader.shader_node import ShaderNode
-from ..shader.shader_node_base_elements import positionView, normalView, uv, vec3, add, sub, mul, dFdx, dFdy, cross, max, dot, normalize, inversesqrt, faceDirection
+from ..shadernode.shader_node import ShaderNode
+from ..shadernode.shader_node_base_elements import positionView, normalView, uv, vec3, add, sub, mul, dFdx, dFdy, cross, max, dot, normalize, inversesqrt, faceDirection
 
 import three
