+++ three/nodes/functions/BSDF/F_Schlick.py
@@ -1,5 +1,5 @@
-from ...shader.shader_node import ShaderNode
-from ...shader.shader_node_base_elements import add, sub, mul, exp2
+from ...shadernode.shader_node import ShaderNode
+from ...shadernode.shader_node_base_elements import add, sub, mul, exp2
 
 def __F_Schlick( inputs ):
     f0 = inputs['f0']
