+++ three/nodes/lighting/ao_node.py
@@ -1,5 +1,5 @@
 from .lighting_node import LightingNode
-from ..shader.shader_node_base_elements import float, add, mul, sub
+from ..shadernode.shader_node_base_elements import float, add, mul, sub
 
 class AONode(LightingNode):
