+++ three/nodes/shadernode/__init__.py
