+++ three/nodes/shadernode/shader_node.py
